Fix some file circular dependencies
