+++ src/components/files/FileActions/FileAction.enum.ts
@@ -1,0 +1,7 @@
+export enum FileAction {
+  Rename = 'rename',
+  Download = 'download',
+  History = 'history',
+  NewVersion = 'new version',
+  Delete = 'delete',
+}
