+++ src/components/files/FileActions/index.ts
@@ -1,4 +1,5 @@
 export * from './FileActions.generated';
 export * from './FileActionsContext';
 export * from './FileActionsMenu';
+export * from './FileAction.enum';
 export * from './getPermittedFileActions';
