+++ src/components/files/FilePreview/PdfPreview.tsx
@@ -1,7 +1,7 @@
 import type { PDFDocumentProxy } from 'pdfjs-dist';
 import React, { FC, useCallback, useState } from 'react';
 import { Document, Page, pdfjs } from 'react-pdf';
-import { useFileActions } from '../FileActions';
+import { useFileActions } from '../FileActions/FileActionsContext';
 import { PreviewerProps } from './FilePreview';
 import { PreviewLoading } from './PreviewLoading';
 import { PreviewPagination } from './PreviewPagination';
