+++ src/components/files/hooks/DownloadFile.graphql
@@ -1,0 +1,17 @@
+fragment FileDownloadUrl on FileNode {
+  id
+  name
+  type
+  ... on File {
+    downloadUrl
+  }
+  ... on FileVersion {
+    downloadUrl
+  }
+}
+
+query FileNodeDownloadUrl($id: ID!) {
+  fileNode(id: $id) {
+    ...FileDownloadUrl
+  }
+}
